Remove double cast for year in date command
From string to int to add one and to string again in the printf. As we
use this for tests being very explicit protects us in case PHP ever changes.

+++ webapp/tests/Unit/Controller/Jury/ImportExportControllerTest.php
@@ -88,7 +88,7 @@ class ImportExportControllerTest extends BaseTestCase
 
     public static function provideContestYamlContents(): Generator
     {
-        $year = date('Y')+1;
+        $year = (int)date('Y')+1;
         $pastYear = date('Y');
         $yaml =<<<HEREDOC
 id: demo
